cleanup: organize files for readability
