+++ WebApp/src/components/moduleComponents/emptyOutput.jsx
@@ -8,9 +8,9 @@
  * Author: Danielle Chang
  ***************************************************************/
 
-import { moduleRegistry } from "../data/defaultModules";
-import { HTTP_STATUS, MODULE_TYPE } from "../utils/utils";
-import { api } from "../utils/utils.js";
+import { moduleRegistry } from "../../data/defaultModules";
+import { HTTP_STATUS, MODULE_TYPE } from "../../utils/utils";
+import { api } from "../../utils/utils.js";
 
 /**
  * @brief The EmptyOutput component represents an output with no audio analysis modules.
