--- WebApp/src/data/modulesData.json
@@ -1,8 +1,0 @@
-{
-  "type": 0,
-  "freqLow": 0,
-  "freqhigh": 0,
-  "minAmpNorm": 1,
-  "frequencyMapping": 0,
-  "params": {}
-}
