+++ WebApp/src/pages/modulesPage.jsx
@@ -10,12 +10,12 @@
  ***************************************************************/
 
 import { useContext, useEffect, useState } from "preact/hooks";
-import AnalysisModule from "../components/analysisModule";
+import AnalysisModule from "../components/moduleComponents/analysisModule";
 import { api, HTTP_STATUS } from "../utils/utils";
 import { AudioSettingsContext } from "../utils/configurations";
-import GlobalSettings from "../components/globalSettings";
-import ConfigManager from "../components/configManager";
-import EmptyOutput from "../components/emptyOutput";
+import GlobalSettings from "../components/moduleComponents/globalSettings";
+import ConfigManager from "../components/moduleComponents/configManager";
+import EmptyOutput from "../components/moduleComponents/emptyOutput";
 
 const ModulesPage = () => {
   // Persistant memory/data
